+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ## [ADD NEW VERSION HERE] - ADD DATE HERE
 ### Fixed
 ### Added
+ - Check for 'rsync' before starting to collect statistics.
 ### Removed
 ### Changed