+++ CHANGELOG.md
@@ -6,12 +6,15 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ## [ADD NEW VERSION HERE] - ADD DATE HERE
 ### Fixed
 ### Added
+### Removed
+### Changed
+
+## [1.0.53] - 2025-04-07
+### Added
  - Check for 'rsync' before starting to collect statistics.
  - Add support for workloads that provide custom lables. Use the '--pipe' option
    for that. The 'stc-wl-cpu-wake-walk' is an example of such a workload.
  - Add hover text to turbostat scatter plots.
-### Removed
-### Changed
 
 ## [1.0.52] - 2025-03-26
 ### Fixed